way: holder 'fee'
1) `par` is reintroduced as the price of sai (ref per sai)
2) `way` is introduced as the rate of change of `par`
3) `coax` is used to adjust `way`

What does `way` actually do? It changes the value of sai in terms of
ref:

- `way > 1` gives `par` increasing over time, giving more ref per sai
- `way < 1` gives `par` decreasing over time, giving less ref per sai

This affects the safety margin of cups, as this is calculated in terms
of ref. This affects `draw`, `wipe`, `free` and `bite`.

`cage` is slightly affected. The argument to `cage` is the cage price,
which is now explicitly in terms of *sai per gem*. `bail` and `cash` are
unaffected.

`boom` and `bust` need to know the skr / sai price, so need to read
`par`.