fee: continuous CDP holder fees, payable on wipe
Introduces a new risk parameter: the `fee`.

This is similar to the `tax`. Income from the `tax` is transferred to
SKR holders via `boom`, whereas the `fee` is paid in GOV, a governance
token, when debt is wiped from a CDP.

The ratio of GOV to SAI is provided by an external oracle `pep`.

Internally, we track a further rate accumulation, `rhi`, which is
calculated the same as `chi`, except including the `fee`. CDPs track
`irk`, which is the debt normalised for `tax` and `fee` (`art` is
normalised only for `tax`).

This allows us to determine the fraction of the debt ascribing only to
`fee`, which is used in `wipe` to determine the amount to be paid in
GOV.

Importantly, unpaid fees *do not* contribute to the `tab`, which is used
to assess CDP risk. On `bite`, unpaid fees are forgotten.