fee: continuous CDP holder fees, payable on wipe
TODO: look for rounding errors, from extra complexity in `wipe`

Introduces a new risk parameter: the `fee`.

This is similar to the `tax`. Income from the `tax` is transferred to
SKR holders via `boom`, whereas the `fee` is paid in GOV, a governance
token, when debt is wiped from a CDP.

The ratio of GOV to SAI is provided by an external oracle `pep`.

Internally, we track a further rate accumulation, `rhi`, which is
calculated the same as `chi`, except including the `fee`. CDPs track the
`rap`, which is the debt normalised for `tax` and `fee` (`art` is
normalised only for `tax`).

This allows us to determine the fraction of the debt ascribing only to
`fee`, which is used in `drip` to calculate the additional sin to mint,
and in `wipe` to determine the fraction of the wiped amount to be paid
in GOV.