Initial Nixification
This adds another way of building the Sai contracts: as a Nix package
with dependencies on the Dappsys Nix packages.

The Nix way of making Solidity packages is still under development, and
using it with the Sai codebase is an experiment I propose.

One consequence is that users need only have Nix itself installed to
build the contracts; they don't need to have "dapp" or any other such
tool installed.

The Git revision and SHA256 in `default.nix` points out the exact
"recipe" for building the contracts, as well as a specific version of
the Dappsys framework.  This is a way to avoid any ambiguity about
compiler versions and such.

The dependency points out a "fork" of Dappsys which (1) merges all ds-*
into a monorepo; (2) has no submodules in Git; and (3) is built without
using the "dapp" tool.  This fork will hopefully become the
main repository.