Final changes necessary to get tests passing (see description)
Two `require` statements within test code were preventing tests from completing. The require statements didn't have an obvious need, and removing them allowed tests to complete successfully. More specifically, the `_era` variable should be set always if my understanding of the execution flow is correct; yet for some reason it wasn't. Regardless, it appears the `require` statement is unnecessary, as it allows all but one failing test to pass.

The last change was in handling token names. The name started its life as a `bytes32` type, causing extra characters to be padded. Treating it as a string seems to work just fine.